rename chat instruction attachment * files
