+++ src/vs/workbench/contrib/chat/browser/attachments/instructionsAttachment/instructionsAttachmentWidget.ts
